+++ CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.12.5] - 2026-01-05
+## [0.12.6] - 2026-01-07
 
 ### Fixed
-- `iter_mut`
+- Attempt to subtract with overflow with reverse iterator: https://github.com/brurucy/indexset/issues/41
+
+[0.12.5] - 2026-01-05
+
+### Fixed
+- `iter_mut` yielding an empty iterator after `.rev` having been called: https://github.com/brurucy/indexset/issues/43
 
 ## [0.12.4] - 2026-01-04
 
@@ -235,7 +240,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `BTreeSet`
 - `BTreeMap`
 
-[Unreleased]: https://github.com/brurucy/indexset/compare/v0.12.4...HEAD
+[Unreleased]: https://github.com/brurucy/indexset/compare/v0.12.6...HEAD
+
+[0.12.6]: https://github.com/brurucy/indexset/releases/tag/v0.12.6
+
+[0.12.5]: https://github.com/brurucy/indexset/releases/tag/v0.12.5
 
 [0.12.4]: https://github.com/brurucy/indexset/releases/tag/v0.12.4